refactor: 빌드 로직 수정
- 빌드 로직이 중복되거나 잘못 사용되고 있던 부분 수정
- 도메인 모듈에 안드로이드 의존성 가지지 않도록 수정
