+++ buildSrc/src/main/kotlin/gradle/plugin/android/AndroidProtobufPlugin.kt
@@ -1,11 +1,9 @@
 package gradle.plugin.android
 
-import com.android.build.api.dsl.LibraryExtension
-import gradle.plugin.configure.configureProtobuf
 import gradle.plugin.extension.getVersionCatalog
+import gradle.plugin.extension.protobufExtension
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.kotlin.dsl.configure
 import org.gradle.kotlin.dsl.dependencies
 
 class AndroidProtobufPlugin : Plugin<Project> {
@@ -15,23 +13,29 @@ class AndroidProtobufPlugin : Plugin<Project> {
             apply("com.google.protobuf")
         }
 
-        extensions.configure<LibraryExtension> {
-            defaultConfig {
-                testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-            }
-
-            testOptions {
-                unitTests.all {
-                    it.useJUnitPlatform()
-                }
-            }
-        }
-
         configureProtobuf()
 
         val libs = getVersionCatalog()
         dependencies {
             "implementation"(libs.findLibrary("protobuf").get())
+        }
+    }
+}
+
+internal fun Project.configureProtobuf() {
+    protobufExtension.apply {
+        protoc {
+            artifact = "com.google.protobuf:protoc:3.25.0"
+        }
+
+        generateProtoTasks {
+            all().forEach { task ->
+                task.builtins {
+                    create("java") {
+                        option("lite")
+                    }
+                }
+            }
         }
     }
 }
