+++ buildSrc/src/main/kotlin/gradle/plugin/configure/KotlinAndroid.kt
@@ -1,10 +1,8 @@
 package gradle.plugin.configure
 
 import gradle.plugin.extension.androidExtension
-import org.gradle.api.JavaVersion
 import org.gradle.api.Project
-import org.gradle.kotlin.dsl.withType
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+import org.jetbrains.kotlin.gradle.dsl.kotlinExtension
 
 internal fun Project.configureKotlinAndroid() {
     androidExtension.apply {
@@ -14,15 +12,8 @@ internal fun Project.configureKotlinAndroid() {
             minSdk = 26
         }
 
-        compileOptions {
-            sourceCompatibility = JavaVersion.VERSION_17
-            targetCompatibility = JavaVersion.VERSION_17
-        }
-    }
-
-    tasks.withType<KotlinCompile>().configureEach {
-        compilerOptions {
-            jvmTarget.set(org.jetbrains.kotlin.gradle.dsl.JvmTarget.JVM_17)
+        kotlinExtension.apply {
+            jvmToolchain(17)
         }
     }
 }
